Add migrations, services, models, Controller the Products

+++ Sales system/Models/BusinessProduct.cs
@@ -8,7 +8,7 @@ namespace Sales_system.Models
     public partial class BusinessProduct
     {
         public long FkBusinessId { get; set; }
-        public int FkProductId { get; set; }
+        public long FkProductId { get; set; }
 
         public virtual Business FkBusiness { get; set; }
         public virtual Product FkProduct { get; set; }
