+++ Sales system/Models/SuppliersProduct.cs
@@ -8,7 +8,7 @@ namespace Sales_system.Models
     public partial class SuppliersProduct
     {
         public long FkSupplierId { get; set; }
-        public int FkProductId { get; set; }
+        public long FkProductId { get; set; }
 
         public virtual Product FkProduct { get; set; }
         public virtual Supplier FkSupplier { get; set; }
